parser map files changed to checker map + checker map approved (check there is no space next to our player)
