--- src/main_map_parser.c
@@ -1,54 +1,0 @@
-#include "cub3d.h"
-
-int	main(int argc, char **argv)
-{
-	(void)argc;
-	(void)argv;
-	char	*valid_map[] =
-	{"        1111111111111111111111111",
-		"        1000000000110000000000001   ",
-		"        1011000001110000000000001",
-		"        1001000000000000000000001",
-		"111111111011000001110000000000001",
-		"100000000011000001110111111111111",
-		"11110111111111011100000010001",
-		"11110111111111011101010010001",
-		"11000000110101011100000010001",
-		"10000000000000001100000010001",
-		"10000000000000001101010010001",
-		"11000001110101011111011110N0111",
-		"11110111 1110101 101111010001",
-		"11111111 1111111 111111111111", NULL};
-
-	char	*invalid_map[] =
-	{"         1000000000110000000000001",
-		"        1111111111111111111111111",
-		"        1011000001110000000000001",
-		"        1001000000000000000000001",
-		"111111111011000001110000000000001",
-		"100000000011000001110111111111111",
-		"11110111111111011100000010001",
-		"11110111111111011101010010001",
-		"11000000110101011100000010001",
-		"10000000000000001100000010001",
-		"10000000000000001101010010001",
-		"11000001110101011111011110N0111",
-		"11110111 1110101 101111010001",
-		"11111111 1111111 111111111111", NULL};
-
-	if (map_is_valid(valid_map) == 1)
-		printf("valid_map = ok\n");
-	else
-		printf("valid_map = ko\n");
-
-	if (map_is_valid(invalid_map) == 1)
-		printf("invalid_map = ok\n");
-	else
-		printf("invalid_map = ko\n");
-
-	// t_all *all;
-	//
-	// init_all(&all);
-	// parser(argc, argv, all);
-	return (0);
-}
